+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fixed and error that prevented payments on orders in the control panel. ([#1362](https://github.com/craftcms/commerce/issues/1362)) 
 - Fixed a bug that caused a redirect back to the Edit Customer page instead of the Edit User page editing an address from the Customer Info tab. ([#1368](https://github.com/craftcms/commerce/issues/1368)) 
 - Hard deleting a product now correctly hard deletes orphaned variants. ([#1186](https://github.com/craftcms/commerce/issues/1186))
+- Fixed a bug when that occurred when selecting all orders on the order index page. ([#1367](https://github.com/craftcms/commerce/issues/1367))
  
 ## 3.1.1 - 2020-04-03